FEAT: 몬스터 데이터 작업

+++ Client/Game/Subsystems/DataSubsystem.h
@@ -7,7 +7,7 @@
 struct MobStats
 {
     int32_t lv;
-    float hp;
+    int32_t hp;
     float speed;
 };
